+++ bin/dsg.dart
@@ -4,6 +4,6 @@ import 'dart:io';
 import 'package:dsg/dsg.dart';
 
 Future main(List<String> arguments) async {
-  final Application application = Application();
+  final application = Application();
   exitCode = await application.run(arguments);
 }